refactor: 💡 improvements on REST endpoints
Refactoring code related to REST endpoints

BREAKING CHANGE: 🧨 `POST /case/list` was replaced on two new endpoints: `GET /cases` and
`GET /user/:id/cases`

+++ src/adonisjs/app/Controllers/Http/v1/CaseController.js
@@ -14,28 +14,13 @@ const uuidv4 = require('uuid/v4');
 
 /** * Resourceful controller for interacting with cases */
 class CaseController {
+  
   /** Show a list of all cases */
   async index({ request, response, }) {
     try {
-
-      let filterBy = request.get().filterBy
-      let filter = request.get().filter
-
-      if (filterBy == null){
-        // let cases = await Case.query().with('versions').fetch()
-        let cases = await Case.all()
-
-        return response.json(cases)
-      }
-
-      if (filterBy == 'user'){
-        let user = await User.find(filter)
-        let cases = await user.contributes_with_cases().fetch()
-
-        return response.json(cases)
-      }
-    } catch (e) {
-      console.log(e)
+      let cases = await Case.all()
+      return response.json(cases)
+    } catch(e){
       return response.status(500).json({ message: e.message })
     }
   }
@@ -105,26 +90,38 @@ class CaseController {
   }
 
   /** * Update case details. PUT or PATCH case/:id */
-  async update({ params, request, response }) {
+  async update({ params, request, response, auth }) {
     try {
+      let logged_user = auth.user.id
+
+      // verify if the loged user is contributor of the given case
+      let sqlQuery = 'select cc.user_id from users u ' +
+                        'left join case_contributors cc on u.id = cc.user_id ' +
+                        'where cc.user_id = ? and cc.case_id = ? and cc.role = 1'
+      let contributor = await Database.raw(sqlQuery, [logged_user, params.id])
+
+      if (contributor == null)
+        return response.status(500).json('you are not contributor of this case')
+
+
       let c = await Case.find(params.id)
 
-      if (c != null){
-        c.title = request.input('title')
-        c.description = request.input('description')
-        c.language = request.input('language')
-        c.domain = request.input('domain')
-        c.specialty = request.input('specialty')
-        c.keywords = request.input('keywords')
-        
-        let cv = new CaseVersion()
-        cv.source = request.input('source')
-        cv.id = await uuidv4()
 
-        await c.versions().save(cv)
-        await c.save() 
-        return response.json(c)
-      } else return response.status(500).json('case not found')
+      if (c != null){
+         c.title = request.input('title')
+         c.description = request.input('description')
+         c.language = request.input('language')
+         c.domain = request.input('domain')
+         c.specialty = request.input('specialty')
+         c.keywords = request.input('keywords')
+          
+         let cv = new CaseVersion()
+         cv.source = request.input('source')
+         cv.id = await uuidv4()
+         await c.versions().save(cv)
+         await c.save() 
+         return response.json(c)
+       } else return response.status(500).json('case not found')
 
     } catch (e) {
       console.log(e)
@@ -140,32 +137,40 @@ class CaseController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-  async destroy({ params, response }) {
+  async destroy({ params, response, auth }) {
     const trx = await Database.beginTransaction()
 
     try {
+      let logged_user = auth.user.id
+
+      // verify if the loged user is owner of the case
+      let sqlQuery = 'select cc.user_id from users u ' +
+                        'left join case_contributors cc on u.id = cc.user_id ' +
+                        'where cc.user_id = ? and cc.case_id = ? and cc.role = 0'
+      let author = await Database.raw(sqlQuery, [logged_user, params.id])
+
+      if (author == null)
+        return response.status(500).json('you are not owner of this case')
+
       let c = await Case.findBy('id', params.id)
       
-        if (c != null){
-                let versions = await c.versions().fetch()
-        let cvs = versions.rows
-        // let case_versions = []
+      if (c != null){
+        let versions = await c.versions().fetch()
+        let versions_rows = versions.rows
 
-        for (let i = 0; i < cvs.length; i++) {
-          let cv = await CaseVersion.findBy('id', cvs[i].id)
+        for (let i = 0; i < versions_rows.length; i++) {
+          let cv = await CaseVersion.findBy('id', versions_rows[i].id)
           cv.delete()
-          // case_versions.push(cvs[i].id)
         }
 
         let contributors = await c.contributors().fetch()
-        let contributors_int = contributors.rows
+        let contributors_rows = contributors.rows
         let contributors_ids = []
 
-        for (let i = 0; i < contributors_int.length; i++) {
-          contributors_ids.push(contributors_int[i].id)
+        for (let i = 0; i < contributors_rows.length; i++) {
+          contributors_ids.push(contributors_rows[i].id)
         }
 
-        // await c.versions().detach(case_versions)
         await c.contributors().detach(contributors_ids)
 
 
@@ -174,7 +179,6 @@ class CaseController {
         trx.commit()
         return response.json(c)
       } else return response.status(500).json('case not found')
-
     } catch (e) {
       console.log(e)
       trx.rollback()
@@ -182,27 +186,34 @@ class CaseController {
     }
   }
 
-  async share({ request, auth, response }) {
+  async share({ request, auth, response }) {  
     try {
-      let loggedUser = auth.user.id
+      let logged_user = auth.user.id
       let {user_id, case_id} = request.post()
 
-      let sqlQuery = 'select c.user_id, c.case_id from users u ' +
-                        'left join contributors c on u.id = c.user_id ' +
-                        'where c.user_id = ? and c.case_id = ?'
-      let contributors = await Database.raw(sqlQuery, [loggedUser, case_id])
+      if (logged_user == user_id){
+        return response.status(500).json('cannot share a case with herself')
+      }
 
-      if (contributors[0] == undefined){
-        return response.json('You are not allowed to share this case.')
+      // get cases in which logged_user is author of
+      let sqlQuery = 'select cc.case_id from users u ' +
+                        'left join case_contributors cc on u.id = cc.user_id ' +
+                        'where cc.user_id = ? and cc.case_id = ? and cc.role = 0'
+      let owned_cases = await Database.raw(sqlQuery, [logged_user, case_id])
+
+      console.log(owned_cases)
+      if (owned_cases[0] == undefined ){
+        return response.status(500).json('you are not owner of this case')
       }
 
       let user = await User.find(user_id)
 
-      await user.cases().attach(case_id, (row) => {
-        row.author = false
+      await user.contributes_with_cases().attach(case_id, (row) => {
+        row.role = 1
       })
 
-      return response.json(contributors[0])
+      return response.json('case successfully shared')
+
     } catch (e) {
       console.log(e)
       return response.status(e.status).json({ message: e.toString() })
